backend_project/requirements.txt
Django==5.0.2
pydantic==1.10.14
minio==7.2.5
gunicorn==21.2.0
jsonpickle==3.0.3
channels==4.0.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9